# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Valuz Agent API
  version: "0.3.0"
  description: Valuz Agent API contract — single source of truth
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/docs:
    get:
      summary: List documents
      operationId: listDocuments
      tags: [docs]
      parameters:
        - name: q
          in: query
          description: Search by filename
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status (queued, processing, ready, failed)
          schema:
            type: string
            enum: [queued, processing, ready, failed]
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/DocumentListItem"
...
  /v1/docs/reindex:
    post:
      summary: Reparse selected documents
      operationId: reindexDocuments
      tags: [docs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_ids]
              properties:
                document_ids:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Reindex task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportTaskResponse"
  /v1/docs/search:
    post:
      summary: Search documents via grep
      operationId: searchDocuments
      tags: [docs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                required: [hits]
                properties:
                  hits:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchHit"
  /v1/docs/tasks/{task_id}:
    get:
      summary: Get import task status
      operationId: getImportTask
      tags: [docs]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportTaskResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/docs/{doc_id}:
    get:
      summary: Get document details
      operationId: getDocument
      tags: [docs]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      summary: Delete a document
      operationId: deleteDocument
      tags: [docs]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
  /v1/docs/{doc_id}/preview:
    get:
      summary: Get document markdown preview
      operationId: getDocumentPreview
      tags: [docs]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Markdown preview content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentPreview"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    DocumentListItem:
      type: object
      required: [id, filename, status, chunk_count, file_size_bytes]
      properties:
        id:
          type: string
        filename:
          type: string
        title:
          type: string
          nullable: true
        status:
          type: string
          enum: [queued, processing, ready, failed]
        chunk_count:
          type: integer
        file_size_bytes:
          type: integer
        mime_type:
          type: string
          nullable: true
        created_at:
          type: integer
          format: int64
          description: "Unix epoch milliseconds (UTC). Format via new Date(ms)."
    ParserAttempt:
      type: object
      required: [plugin_id, error, occurred_at, ok]
      properties:
        plugin_id:
          type: string
          description: Identifier of the parser plugin that was invoked
        error:
          type: string
          description: Error message if the attempt failed; empty string on success
        occurred_at:
          type: string
          format: date-time
          description: ISO-8601 timestamp of the attempt
        ok:
          type: boolean
          description: True if this attempt succeeded; false for failed/fallback attempts
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/DocumentListItem"
        - type: object
          properties:
            original_path:
              type: string
              nullable: true
            managed_path:
              type: string
              nullable: true
            parser_mode:
              type: string
              nullable: true
            docs_runtime_id:
              type: string
              nullable: true
            last_error_code:
              type: string
              nullable: true
            last_error_message:
              type: string
              nullable: true
            parser_attempts:
              type: array
              items:
                $ref: "#/components/schemas/ParserAttempt"
              description: Ordered list of parse attempts for this document (oldest first)
    DocumentPreview:
      type: object
      required: [document_id, markdown]
      properties:
        document_id:
          type: string
        markdown:
          type: string
          description: Parsed markdown content
    ImportTaskResponse:
      type: object
      required: [task_id, status, total_items, processed_items, failed_items]
      properties:
        task_id:
          type: string
        task_type:
          type: string
          enum: [import_files, import_folder, reindex]
        status:
          type: string
          enum: [queued, processing, completed, failed]
        total_items:
          type: integer
        processed_items:
          type: integer
        failed_items:
          type: integer
        project_id:
          type: string
          nullable: true
        created_at:
          type: integer
          format: int64
          description: "Unix epoch milliseconds (UTC). Format via new Date(ms)."
    SearchRequest:
      type: object
      required: [query, project_id]
      properties:
        query:
          type: string
        project_id:
          type: string
        top_k:
          type: integer
          default: 5
    SearchHit:
      type: object
      required: [document_id, filename, score, snippet]
      properties:
        document_id:
          type: string
        filename:
          type: string
        score:
          type: number
          format: float
        snippet:
          type: string
        page_ref:
          type: string
          nullable: true
        chunk_ref:
          type: string
          nullable: true
...